requirements.txt
openai
python-dotenv
Flask
pytest-playwright